chore: reorganized project structure and removed files relating to TUI implementation
