--- src/github_feed/css/tui.tcss
@@ -1,31 +1,0 @@
-Screen {
-    layout: vertical;
-    align: center middle;
-}
-
-EnvVarPanel {
-    height: auto;
-    width: 50%;
-    content-align: right middle;
-    padding-right: 5;
-}
-
-#metadataPanel {
-    height: auto;
-    width: 50%;
-    content-align: left middle;
-    padding-left: 5;
-}
-
-.row {
-    height: 1fr;
-    align: center middle;
-}
-
-#checkReleases {
-    margin-right: 1;
-}
-
-#checkStarred {
-    margin-left: 1;
-}
