+++ src/github_feed/lib/utils.py
@@ -2,7 +2,7 @@ import pathlib
 
 import urllib3
 
-from github_feed.models import LinkHeader, Repository
+from github_feed.lib.models import LinkHeader, Repository
 from github_feed.sql.models import Repository as SqlRepository
